Update remaining gems
These changes were brought in with Cursor
iterating. I don't think there is much more
which can be done as the Aruba gem is pinned
to much older versions of cucumber tooling,
and while aruba is under active development,
there is not indication of updating the
cucumber version dependency. A longer temr
goal would be to remove the aruba dependency
completely.